+++ Fabric/src/main/java/com/almostreliable/lootjs/fabric/kubejs/LootModificationFabricEventJS.java
@@ -1,6 +1,0 @@
-package com.almostreliable.lootjs.fabric.kubejs;
-
-import com.almostreliable.lootjs.kube.LootModificationEventJS;
-
-public class LootModificationFabricEventJS extends LootModificationEventJS {
-}